Add a download button
The download button always downloads, even if there are other save
methods active.

It's hidden by default, but can be easily accessed or made visible
in the same way as the other buttons in the side bar tools list.

The concept is based on the demo [1] which is based on Justin_H's
original example [2] but I did change it to follow the format of the
other core buttons so it can be hidden and shown consistently.

[1] https://simon.tiddlyhost.com/#Download%20button%20demo
[2] https://talk.tiddlywiki.org/t/help-creating-a-button-to-save-locally-and-button-to-save-to-tidddlyhost/3824