Switch to light default color scheme
Also make it the same for both variants, since I don't want people
to be choosing one or the other based on the color.

I'm not sure I prefer the light color schemes personally, but the
new palettes are really good and I feel like trying this as a
default.