Store build timestamp in separate tiddler
Because the modified field I was using will be updated if the user
changes the tiddler, it's not a suitable place to record a build
timestamp. Let's put it somewhere else.

(I considered switching to using the "created" field instead of
"modified" for those two tiddlers, but I suspect I would still need
a modified field to have it show up in the "Recent" tab, so I decide
it's not worth the effort to change it.)